Revert "(merge) Replaces uint32_t with double when working with scales"
This reverts commit de83a6598d75ecd8e205cd367cc706a54b8ceec7.

Reason for revert: A parent CL (36412d77e4e8a7d5b6d06d8f2861412c007c65ef) needs to be reverted and this cl is in conflict of that. 

Original change's description:
> (merge) Replaces uint32_t with double when working with scales
> 
> Merge to M67
> Zoom factors are double, but the code is using uint32_t to process them.
> This patch fixes it.
> 
> Output on wayland info client:
> OUTPUT0:
> 
>   connection:          internal
>   device scale factor: 1.000
> 
>   geometry:
>     x:                 2
>     y:                 24
>     physical width:    443 mm
>     physical height:   264 mm
>     subpixel:          unknown
>     make:              unknown
>     model:             unknown
>     transform:         normal
> 
>   modes:
>     1676x998:          60.00 Hz      current       preferred
> 
>   scales:
>     0.500:
>     0.750:
>     1.000:             current       preferred
>     1.250:
>     1.500:
>     1.750:
>     2.000:
>     2.250:
>     2.500:
> 
> Bug: 836782
> Component: Wayland, aura shell, scales
> Reviewed-on: https://chromium-review.googlesource.com/1028217
> Commit-Queue: Malay Keshav <malaykeshav@chromium.org>
> Cr-Original-Commit-Position: refs/heads/master@{#553586}(cherry picked from commit e24cbe8a0ed6a7b9d96dc34560c8909ad75d2764)
> Reviewed-on: https://chromium-review.googlesource.com/1030771
> Cr-Commit-Position: refs/branch-heads/3396@{#328}
> Cr-Branched-From: 9ef2aa869bc7bc0c089e255d698cca6e47d6b038-refs/heads/master@{#550428}

TBR=malaykeshav@chromium.org

No-Presubmit: true
No-Tree-Checks: true
No-Try: true
Bug: 836782
Reviewed-on: https://chromium-review.googlesource.com/1030576
Cr-Commit-Position: refs/branch-heads/3396@{#330}
Cr-Branched-From: 9ef2aa869bc7bc0c089e255d698cca6e47d6b038-refs/heads/master@{#550428}

+++ components/exo/wayland/server.cc
@@ -2872,9 +2872,10 @@ class AuraOutput : public WaylandDisplayObserver::ScaleObserver {
         bool rv = display_manager->GetActiveModeForDisplayId(display.id(),
                                                              &active_mode);
         DCHECK(rv);
-        for (double zoom_factor : display::GetDisplayZoomFactors(active_mode)) {
+        for (uint32_t zoom_factor :
+             display::GetDisplayZoomFactors(active_mode)) {
           uint32_t flags = 0;
-          if (zoom_factor == 1.0)
+          if (zoom_factor == 1.f)
             flags |= ZAURA_OUTPUT_SCALE_PROPERTY_PREFERRED;
           if (display_info.zoom_factor() == zoom_factor)
             flags |= ZAURA_OUTPUT_SCALE_PROPERTY_CURRENT;
